Move the Frame and the FrameNames class in Frame.* files
