+++ src/OptimalControlUtilities/CMakeLists.txt
@@ -19,6 +19,7 @@ if(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_OptimalControlUtilities)
     src/VariableHandler.cpp
     src/FloatingBaseMultiBodyDynamicsElements.cpp
     src/ZMPElements.cpp
+    src/Frame.cpp
     )
 
   # set hpp files
@@ -34,6 +35,7 @@ if(BIPEDAL_LOCOMOTION_CONTROLLERS_COMPILE_OptimalControlUtilities)
     include/BipedalLocomotionControllers/OptimalControlUtilities/FloatingBaseMultiBodyDynamicsElements.h
     include/BipedalLocomotionControllers/OptimalControlUtilities/ZMPElements.h
     include/BipedalLocomotionControllers/OptimalControlUtilities/OptimizationProblemElements.h
+    include/BipedalLocomotionControllers/OptimalControlUtilities/Frame.h
     )
 
   # add an executable to the project using the specified source files.
