+++ src/OptimalControlUtilities/include/BipedalLocomotionControllers/OptimalControlUtilities/ControlProblemElements.h
@@ -22,43 +22,13 @@
 #include <iDynTree/Core/VectorDynSize.h>
 #include <iDynTree/KinDynComputations.h>
 
+#include <BipedalLocomotionControllers/OptimalControlUtilities/Frame.h>
+
 namespace BipedalLocomotionControllers
 {
 
 namespace OptimalControlUtilities
 {
-
-/**
- * The struct frame defines a Frame in a ControlProblemElement
- */
-struct Frame
-{
-    iDynTree::FrameIndex indexInModel; /**< Index of the frame in the Model */
-
-    /** Index and size of the element in the ControlProblemElement */
-    iDynTree::IndexRange indexRangeInElement{iDynTree::IndexRange::InvalidRange()};
-};
-
-/**
- * FrameNames wraps a std::pair<std::string, std::string>. It contains the label
- * associated to the frame in the VariableHandler end the name of the frame in the model
- */
-class FrameNames : public std::pair<std::string, std::string>
-{
-    // obscure first and second attribute
-    using std::pair<std::string, std::string>::first;
-    using std::pair<std::string, std::string>::second;
-
-public:
-
-    // use the std::pair constructors
-    using std::pair<std::string, std::string>::pair;
-
-    std::string& label();
-    const std::string& label() const;
-    std::string& nameInModel();
-    const std::string& nameInModel() const;
-};
 
 /**
  * ControlProblemElement describes a general control problem element. The
